requirements.txt
tiktoken
faiss-cpu
langchain
fitz
PyPDF2
python-dotenv
streamlit
openai